% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkboxInputs.R
\name{checkboxInputs}
\alias{checkboxInput_cell}
\alias{checkboxInput_row}
\alias{checkboxInputs}
\title{Tabluar Checkbox Control Elements}
\usage{
checkboxInput_cell(inputId, label = "", value = FALSE, width = "",
  leftLabel = FALSE)

checkboxInput_row(inputId, label = "", value = FALSE, width = "",
  leftLabel = FALSE)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control.  Defaults to \code{""} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE})}

\item{width}{The width of the input, e.g. \code{'400px'} or '100\%'. Defaults to
\code{""}.}

\item{leftLabel}{Should the label column be the left column or the right column.
  \code{TRUE} puts the label on the left column.}
}
\value{
\code{checkboxInput_cell} returns a character vector of HTML code, one
  element for each \code{inputId}.

\code{checkboxInput_row} returns a data frame with two columns and one row for
  each \code{inputId}.
}
\description{
Produce the HTML code for a cell with a checkbox or a row of a two
  column table iwth a checkbox.
}
\examples{
\dontrun{
options(pixiedust_print_method = "html")
shinyApp(
  ui = shinyUI(
    wellPanel(uiOutput("sampleUI"))
  ),

  server = shinyServer(function(input, output){
    output$sampleUI <-
      renderText({
        checkboxInput_row(inputId = paste0("car", 1:nrow(mtcars)),
                          label = rownames(mtcars)) \%>\%
                          dust() \%>\%
                          print(asis = FALSE)
      })
  })
)
}
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link[shiny]{checkboxInput}}
}

