% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateInputs.R
\name{dateInput_cell}
\alias{dateInput_cell}
\alias{dateInput_row}
\title{Create Tabular Date Inputs}
\usage{
dateInput_cell(
  inputId,
  label = "",
  value = NULL,
  min = NULL,
  max = NULL,
  format = "yyyy-mm-dd",
  startview = "month",
  weekstart = 0,
  language = "en",
  width = "",
  emptyValueNoDefault = TRUE,
  disabled = FALSE,
  hidden = FALSE
)

dateInput_row(
  inputId,
  label = "",
  value = NULL,
  min = NULL,
  max = NULL,
  format = "yyyy-mm-dd",
  startview = "month",
  weekstart = 0,
  language = "en",
  width = "",
  leftLabel = TRUE,
  disabled = FALSE,
  hidden = FALSE
)
}
\arguments{
\item{inputId}{the input slot that will be used to acces the value}

\item{label}{Display label for the control, or \code{""} for no label.}

\item{value}{The starting date. Either a date object, or a string in 
\code{yyyy-mm-dd} format.  If \code{NULL} (the default), will use the
current date in the client's time zone.}

\item{min}{the minimum allowed date.  Either a date object, or a string in 
\code{yyyy-mm-dd} format.}

\item{max}{the maximum allowed date.  Either a date object, or a string in 
\code{yyyy-mm-dd} format.}

\item{format}{The format for the date to display in the browser.  Defaults
to \code{"yyyy-mm-dd"}}

\item{startview}{The date range shown when the input object is first
clicked.  Can be \code{"month"} (the default), \code{"year"}, or 
\code{"decade"}.}

\item{weekstart}{Which day is the start of the week.  Should be an integer
from 0 (Sunday) to 6 (Saturday)}

\item{width}{The width of the input, eg \code{'400px'} or \code{'100\%'}}

\item{The}{language used for month and day names.  Default is \code{"en"}.
See \code{\link[shiny]{dateInput}} for other valid options.}
}
\value{
\code{dateInput_cell} returns a character vector the length of \code{inputId}.

\code{dateInput_row} returns a data frame with two columns and a row for every 
element in \code{inputId}
}
\description{
Create tabular text inputs that, when clicked on, bring up a calendar
  that the user can click on to select dates.
}
\details{
For more about \code{format} strings, see the documentation for 
  \code{\link[shiny]{dateInput}}
}
\examples{
\dontrun{
options(pixiedust_print_method = "html")
shinyApp(
  ui = shinyUI(
    navbarPage(title = "Example",
      tabsetPanel(
        tabPanel(title = "Dates",
          uiOutput("sampleUI")
        )
      )
    )
  ),
  
  server = shinyServer(function(input, output){
    output$sampleUI <- 
      renderText({ 
        dateInput_row(inputId = paste0("car", 1:nrow(mtcars)),
                      label = rownames(mtcars)) \%>\%
        dust() \%>\%
        print(asis = FALSE)
      })
      
  })
)
}

}
\seealso{
\code{\link[shiny]{dateInput}}, \code{\link[shiny]{dateRangeInput}}
}
\author{
Benjamin Nutter
}
