% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadButtons.R
\name{downloadButtons}
\alias{downloadButtons}
\alias{downloadButton_cell}
\alias{downloadButton_row}
\title{Action Buttons and Links}
\usage{
downloadButton_cell(
  outputId,
  label = "Download",
  class = NULL,
  leftLabel = TRUE,
  labelButton = TRUE,
  disabled = FALSE,
  hidden = FALSE
)

downloadButton_row(
  outputId,
  label = "Download",
  class = NULL,
  leftLabel = TRUE,
  labelButton = TRUE,
  disabled = FALSE,
  hidden = FALSE
)
}
\arguments{
\item{label}{The contents of the button or link-usually a text label, but
you could also use any other HTML, like an image.}

\item{class}{Additional CSS classes to apply to the tag, if any.  Currently
not implemented.}

\item{leftLabel}{Should the label column be the left column or the right column.
\code{TRUE} puts the label on the left column.  This only takes effect when
\code{labelButton = FALSE}.}

\item{labelButton}{Should the label be placed on the button? If \code{FALSE}, the 
label text appears to the side of the button and the button itself is blank.}

\item{inputId}{The \code{input} slot that will be used to access the value.}
}
\description{
Create action buttons in tabular input panels.
}
\examples{
\dontrun{
options(pixiedust_print_method = "html")

#* Example of Download Buttons
shinyApp(
  ui = shinyUI(
    wellPanel(uiOutput("sampleUI"))
  ),
  
  server = shinyServer(function(input, output){
    output$sampleUI <- 
      renderText({ 
        downloadButton_row(inputId = paste0("car", 1:nrow(mtcars)),
                          label = rownames(mtcars),
                          labelButton = c(TRUE, FALSE)) \%>\%
                          dust() \%>\%
                          print(asis = FALSE)
      })
  })
)



#' #* Example of Download Links
shinyApp(
  ui = shinyUI(
    wellPanel(uiOutput("sampleUI"))
  ),
  
  server = shinyServer(function(input, output){
    output$sampleUI <- 
      renderText({ 
        data.frame(
          actionLink = 
            downloadLink_cell(inputId = paste0("car", 1:nrow(mtcars)),
                            label = rownames(mtcars)),
          stringsAsFactors = FALSE) \%>\%
        dust() \%>\%
        print(asis = FALSE)
      })
  })
)
}
}
\seealso{
\code{\link[shiny]{downloadButton}}
}
\author{
Benjamin Nutter
}
