% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/launchExample.R
\name{launchExample}
\alias{launchExample}
\title{Launch an Example Shiny Application}
\usage{
launchExample(example)
}
\arguments{
\item{example}{Character string giving the title of the example application to be
  viewed.}
}
\description{
\code{shinydust} contains several examples of applications that illustrate
 the functionality of tabular panels.  These can be launched quickly by providing the
 name of the example you wish to view.
}
\details{
\tabular{
  Example Name \tab Description \cr
  \code{"textInputTable"} \tab A tabular panel with text input fields \cr
}
}
\examples{
\dontrun{
  launchExample("textInputTable")
}
}
\author{
Benjamin Nutter
}

