% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numericInputs.R
\name{numericInputs}
\alias{numericInput_cell}
\alias{numericInput_row}
\alias{numericInputs}
\title{Tabular Numeric Input Controls}
\usage{
numericInput_cell(inputId, label = "", value, min = NA, max = NA,
  step = NA, width = "")

numericInput_row(inputId, label = "", value, min = NA, max = NA,
  step = NA, width = "", leftLabel = TRUE)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value.}

\item{max}{Maximum allowed value.}

\item{step}{Interval to use when stepping between min and max}

\item{width}{The width of the input, e.g. \code{'400px'} or \code{'100\%'}}

\item{leftLabel}{Should the label column be the left column or the right column.
  \code{TRUE} puts the label on the left column.  This only takes effect when
  \code{labelButton = FALSE}.}
}
\value{
\code{numericInput_cell} returns a character vector the length of \code{inputId}

\code{numericInput_row} returns a data frame with two columns and a row for each element of
  \code{inputId}
}
\description{
Create an input control for entry of numeric values
}
\examples{
\dontrun{
options(pixiedust_print_method = "html")
shinyApp(
  ui = shinyUI(
    wellPanel(uiOutput("sampleUI"))
  ),

  server = shinyServer(function(input, output){
    output$sampleUI <-
      renderText({
        numericInput_row(inputId = paste0("car", 1:nrow(mtcars)),
                         label = rownames(mtcars),
                         value = 50,
                         min = 0,
                         max = 100,
                         step = 10) \%>\%
                         dust() \%>\%
                         print(asis = FALSE)
      })
  })
)
}
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link[shiny]{numericInput}}
}

