% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwordInputs.R
\name{passwordInputs}
\alias{passwordInputs}
\alias{passwordInput_cell}
\alias{passwordInput_row}
\title{Tabular Password Input Controls}
\usage{
passwordInput_cell(
  inputId,
  label = "",
  value = "",
  width = "",
  disabled = FALSE,
  hidden = FALSE
)

passwordInput_row(
  inputId,
  label,
  value = "",
  width = "",
  leftLabel = TRUE,
  disabled = FALSE,
  hidden = FALSE
)
}
\arguments{
\item{inputId}{the \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or "" for no label.}

\item{value}{Initial value}

\item{width}{The width of the input, e.g. \code{'400px'} or \code{'100\%'}}

\item{hidden}{A logical vector specifying which controls should be hidden 
using \code{shinyjs}. Will be recycled if necessary.}

\item{diabled}{A logical vector specifying which controls should be disabled
using \code{shinyjs}. Will be recycled if necessary.}

\item{leftlabel}{Should the label column be the left column or the right 
column.  \code{TRUE} puts the label on the left column.}
}
\description{
Create a table of input controls for entry of passwords.
}
\author{
Benjamin Nutter
}
