% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInputs.R
\name{selectInputs}
\alias{selectInputs}
\alias{selectInput_cell}
\alias{selectInput_row}
\title{Tabular Select List Input Controls}
\usage{
selectInput_cell(
  inputId,
  label = "",
  choices,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE,
  width = "",
  size = NULL,
  disabled = FALSE,
  hidden = FALSE
)

selectInput_row(
  inputId,
  label = "",
  choices,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE,
  width = "",
  size = NULL,
  leftLabel = TRUE,
  disabled = FALSE,
  hidden = FALSE
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value}

\item{label}{Display label for the control, or \code{""} for no label.}

\item{choices}{List of values to select from. If the elements of the list are named, 
then the name rather than the value is displayed to the user.}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}). 
If not specified then defaults to the first value for single-select lists and no 
values for multiple select lists.}

\item{multiple}{Is selection of multiple items allowed?}

\item{selectize}{Whether to use selectize.js or not.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'}}

\item{size}{Number of items to show in the selection box; a larger number will result in
a taller box.  Not compatible with \code{selectize=TRUE}. Normally, when 
\code{multiple=FALSE}, a select input will be a drop-down list, but when \code{size}
is set, it will be a box instead.}

\item{leftLabel}{Should the label column be the left column or the right column.
\code{TRUE} puts the label on the left column.}

\item{...}{Arguments passed to \code{selectInput()}}
}
\value{
\code{selectInput_cell} returns a character vector the length of \code{inputId}.

\code{selectInput_row} returns a data frame with two columns and a row for every 
element in \code{inputId}
}
\description{
Create a select list that can be used to choose a single or multiple items
  from a list of values.
}
\examples{
\dontrun{
options(pixiedust_print_method = "html")
shinyApp(
  ui = shinyUI(
    wellPanel(uiOutput("sampleUI"))
  ),
  
  server = shinyServer(function(input, output){
    output$sampleUI <- 
      renderText({ 
        selectInput_row(inputId = paste0("car", 1:nrow(mtcars)),
                        label = rownames(mtcars),
                        choices = LETTERS[1:5]) \%>\%
        dust() \%>\%
        print(asis = FALSE)
      })
  })
)
}

}
\author{
Benjamin Nutter
}
