% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAreaInputs.R
\name{textAreaInputs}
\alias{textAreaInputs}
\alias{textArea_cell}
\alias{textArea_row}
\title{Create Tabular Text Area Input Controls}
\usage{
textArea_cell(
  inputId,
  label = "",
  value = "",
  leftLabel = TRUE,
  ...,
  disabled = FALSE,
  hidden = FALSE,
  readonly = FALSE
)

textArea_row(
  inputId,
  label = "",
  value = "",
  leftLabel = TRUE,
  ...,
  disabled = FALSE,
  hidden = FALSE
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{""} for no label.}

\item{value}{The initial text to be shown in the text box.}

\item{leftLabel}{Should the label column be the left column or the right column.
\code{TRUE} puts the label on the left column.}

\item{...}{Additional named arguments to paste into the text area tag. Some recommended
tags: \code{rows} (determines how many lines of text are displayed in the box);
\code{cols} The width of the text box. Unnamed arguments will be ignored.}
}
\value{
\code{textArea_cell} returns a character vector the length of \code{inputId}.

\code{textArea_row} returns a data frame with two columns and a row for every 
element in \code{inputId}
}
\description{
Create text area inputs, similar to \code{textInput}, but may be 
  sized as a box to allow for larger (paragraph) inputs.
}
\examples{
\dontrun{
options(pixiedust_print_method = "html")
shinyApp(
  ui = shinyUI(
    wellPanel(uiOutput("sampleUI"))
  ),
  
  server = shinyServer(function(input, output){
    output$sampleUI <- 
      renderText({ 
        textArea_row(inputId = paste0("car", 1:nrow(mtcars)),
                     label = rownames(mtcars),
                     rows = 3,
                     cols = 40) \%>\%
        dust() \%>\%
        print(asis = FALSE)
      })
  })
)
}

}
\author{
Benjamin Nutter
}
