% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInputs.R
\name{textInputs}
\alias{textInputs}
\alias{textInput_cell}
\alias{textInput_row}
\title{Tabular Text Input Controls}
\usage{
textInput_cell(
  inputId,
  label,
  value = "",
  width = "",
  disabled = FALSE,
  hidden = FALSE,
  placeholder = NA
)

textInput_row(
  inputId,
  label,
  value = "",
  width = "",
  leftLabel = TRUE,
  disabled = FALSE,
  hidden = FALSE,
  placeholder = NA
)
}
\arguments{
\item{inputId}{the \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{""} for no label.}

\item{value}{Initial value}

\item{width}{The width of the input, e.g. \code{'400px'} or \code{'100\%'}}

\item{leftLabel}{Should the label column be the left column or the right column.
\code{TRUE} puts the label on the left column.}
}
\value{
\code{textInput_cell} returns a character vector the length of \code{inputId}.

\code{textInput_row} returns a data frame with two columns and a row for every 
element in \code{inputId}
}
\description{
Create a table of input controls for entry of unstructured text values
}
\examples{
\dontrun{
options(pixiedust_print_method = "html")
shinyApp(
  ui = shinyUI(
    wellPanel(uiOutput("sampleUI"))
  ),
  
  server = shinyServer(function(input, output){
    output$sampleUI <- 
      renderText({ 
        textInput_row(inputId = paste0("car", 1:nrow(mtcars)),
                        label = rownames(mtcars)) \%>\%
        dust() \%>\%
        print(asis = FALSE)
      })
  })
)
}

}
\seealso{
\code{\link[shiny]{textInput}}
}
\author{
Benjamin Nutter
}
